chore: move version and terminal to internal
The version and terminal packages are meant to support
the complytime CLI and not be part of the public API
