+++ internal/manager/plugin.go
@@ -1,0 +1,1 @@
+package plugins
