+++ internal/terminal/terminal_test.go
